requirements.txt
langchain-openai
langchain
openai
python-dotenv
langchain-community
tiktoken
langchainhub
chromadb
beautifulsoup4
Wikipedia
pypdf
streamlit